[BE] feature: backend Integration

+++ backend/business/src/main/java/mpti/domain/opinion/api/request/CreateReviewRequest.java
@@ -13,7 +13,7 @@ public class CreateReviewRequest {
 
     private String targetName;
 
-    private Double star;
+    private int star;
 
     private String memo;
